+++ packages/ironbean/src/containerStorage.ts
@@ -1,9 +1,11 @@
 import {CollectionToken, ComponentContainer, Container, IPlugin, PluginToken, take} from "./internals";
+import {StorageMode} from "./storageMode";
 
 class ContainerStorage {
     public container: Container | null = null;
     public currentComponentContainer: ComponentContainer | undefined;
     public currentContainer: Container | undefined;
+    public mode: StorageMode = StorageMode.None;
     private _plugins: IPlugin[] = [];
 
     get plugins(): IPlugin[] {
@@ -11,16 +13,35 @@ class ContainerStorage {
     }
 
     reloadPlugins(): void {
-        this._plugins = this.container?.getBean(CollectionToken.create(PluginToken)) as IPlugin[] ?? [];
+        this._plugins = this.resolvePlugins();
     }
 
-    getBaseContainer(): Container {
+    private resolvePlugins() {
+        if (this.mode === StorageMode.Prototype) {
+            const container = new Container();
+            container.init();
+            return container.getBean(CollectionToken.create(PluginToken)) as IPlugin[];
+        }
+        return this.container?.getBean(CollectionToken.create(PluginToken)) as IPlugin[] ?? [];
+    }
+
+    getOrCreateBaseContainer(): Container {
         if (!this.container) {
-            this.container = new Container();
-            this.initContainer(this.container);
+            if (this.mode === StorageMode.Prototype) {
+                throw new Error("Create global container is not allowed for PROTOTYPE mode.")
+            }
+            this.mode = StorageMode.Singleton;
+            this.container = this.createBaseContainer();
         }
 
         return this.container;
+    }
+
+    createBaseContainer(): Container {
+        const container = new Container();
+        this.initContainer(container);
+
+        return container;
     }
 
     initContainer(container: Container) {