+++ packages/ironbean/src/decorators.ts
@@ -1,6 +1,6 @@
 import "reflect-metadata";
 import {
-    ApplicationContext,
+    ApplicationContext, ClassComponent,
     Component,
     ComponentType,
     constants,
@@ -19,7 +19,7 @@ export function component(ClassOrType: Class | ComponentType): any {
     let componentType = ComponentType.Singleton;
     function decorator(Class: any): any {
         Reflect.defineMetadata(constants.component, true, Class);
-        Reflect.defineMetadata(constants.componentType, componentType, Class);
+        Component.create(Class).setType(componentType);
 
         return Class;
     }
@@ -37,7 +37,7 @@ export function component(ClassOrType: Class | ComponentType): any {
 
 export function scope(scope: Scope): ClassDecorator {
     function decorator(Class: any): any {
-        Reflect.defineMetadata(constants.scope, scope, Class);
+        (Component.create(Class) as ClassComponent<any>).setScope(scope)
 
         return Class;
     }
@@ -46,13 +46,23 @@ export function scope(scope: Scope): ClassDecorator {
 }
 
 export function type<T>(key: DependencyToken<T>|(() => Dependency<T>)) {
-    return function(target: any, propertyName: string | symbol, parameterIndex?: number) {
+    return function(target: any, propertyName?: string | symbol, parameterIndex?: number) {
         if (typeof parameterIndex !== "number") {
-            Reflect.defineMetadata(constants.types, key, target, propertyName);
+            if (propertyName !== undefined) {
+                Reflect.defineMetadata(constants.types, key, target, propertyName);
+            } else {
+                Reflect.defineMetadata(constants.types, key, target);
+            }
         } else {
-            const methodParameters: Object[] = Reflect.getOwnMetadata(constants.types, target, propertyName) || [];
-            methodParameters[parameterIndex] = key;
-            Reflect.defineMetadata(constants.types, methodParameters, target, propertyName);
+            if (propertyName !== undefined) {
+                const methodParameters: Object[] = Reflect.getOwnMetadata(constants.types, target, propertyName) || [];
+                methodParameters[parameterIndex] = key;
+                Reflect.defineMetadata(constants.types, methodParameters, target, propertyName);
+            } else {
+                const methodParameters: Object[] = Reflect.getOwnMetadata(constants.types, target) || [];
+                methodParameters[parameterIndex] = key;
+                Reflect.defineMetadata(constants.types, methodParameters, target);
+            }
         }
     }
 }